improve error handling, add return codes to queries

+++ engines/hive/queries/q05/run.sh
@@ -23,7 +23,7 @@ query_run_main_method () {
 	#step 5.  hive && hdfs 		:	cleanup.sql && hadoop fs rm MH
 
 	HDFS_RESULT_FILE="${RESULT_DIR}/logRegResult.txt"
-
+	RETURN_CODE=0
 	if [[ -z "$DEBUG_QUERY_PART" || $DEBUG_QUERY_PART -eq 1 ]] ; then
 		echo "========================="
 		echo "$QUERY_NAME Step 1/3: Executing hive queries"
@@ -31,7 +31,8 @@ query_run_main_method () {
 		echo "========================="
 		# Write input for k-means into ctable
 		runCmdWithErrorCheck runEngineCmd -f "$QUERY_SCRIPT"
-	fi
+		RETURN_CODE=$?
+		if [[ $RETURN_CODE -ne 0 ]] ;  then return $RETURN_CODE; fi
 
 	if [[ -z "$DEBUG_QUERY_PART" || $DEBUG_QUERY_PART -eq 2 ]] ; then
 	
@@ -66,7 +67,9 @@ query_run_main_method () {
 		echo "-------------------------"
 	
 		runCmdWithErrorCheck mahout trainlogistic --input "$TMP_LOG_REG_IN_FILE" --output "$TMP_LOG_REG_MODEL_FILE" --target c_customer_sk --categories 2 --predictors college_education male label --types n n n --passes 20 --features 20 --rate 1 --lambda 0.5	
-
+		RETURN_CODE=$?
+		if [[ $RETURN_CODE -ne 0 ]] ;  then return $RETURN_CODE; fi
+		
 		echo "-------------------------"
 		echo "$QUERY_NAME Step 2/3 Part 3: Calculating Logistic Regression"
 		echo "Command: " mahout runlogistic --input "$TMP_LOG_REG_IN_FILE" --model "$TMP_LOG_REG_MODEL_FILE" --auc --confusion --quiet 
@@ -74,7 +77,9 @@ query_run_main_method () {
 		echo "-------------------------"
 
 		runCmdWithErrorCheck mahout runlogistic --input "$TMP_LOG_REG_IN_FILE" --model "$TMP_LOG_REG_MODEL_FILE" --auc --confusion --quiet 2> /dev/null | grep -A 3 "AUC =" | hadoop fs -copyFromLocal -f - "$HDFS_RESULT_FILE"
-
+		RETURN_CODE=$?
+		if [[ $RETURN_CODE -ne 0 ]] ;  then return $RETURN_CODE; fi
+		
 		echo "-------------------------"
 		echo "$QUERY_NAME Step 2/3 Part 4: Cleanup tmp files"
 		echo "-------------------------"
@@ -88,10 +93,17 @@ query_run_main_method () {
 		echo "========================="
 	
 		runCmdWithErrorCheck runEngineCmd -f "${QUERY_DIR}/cleanup.sql"
+		RETURN_CODE=$?
+		if [[ $RETURN_CODE -ne 0 ]] ;  then return $RETURN_CODE; fi
+		
 		runCmdWithErrorCheck hadoop fs -rm -r -f "$TEMP_DIR"
+		RETURN_CODE=$?
+		if [[ $RETURN_CODE -ne 0 ]] ;  then return $RETURN_CODE; fi
+		
 	fi
 }
 
 query_run_clean_method () {
 	runCmdWithErrorCheck runEngineCmd -e "DROP TABLE IF EXISTS $TEMP_TABLE; DROP TABLE IF EXISTS $RESULT_TABLE;"
+	return $?
 }
